update info on setting up cryptpad for version v5.4.0
Version 5.4.0 marks the Debowerification of CryptPad. What this
essentially means for the documentation is that the way to setup a dev
instance of CryptPad has changed. Now, a user is required to run an npm
command as npm is used instead of bower for client components.

This commit basically adds new instructions and modifies only ones in
./dev_guide/setup.rst